Improve crypto implementation (backwards compatibility added) (#130)
* Improve Crypto Implementation

Reworking the encryption implementation to be better aligned with best practices. No more hard-coded IV. Utilize PBKDF2 with random salt and IV.

* Added backwards compatibility for legacy encryption

Added logic to first attempt decryption utilizing the legacy method, if successfully decrypted, it updates the encryption immediately to the new SecureAES method. This adds backwards compatibility and migrates the legacy format to the new implementation.

* Flip default decryption method

Modified to make the new SecureAES implementation the default/first method used for decryption and to fallback onto legacy decryption